% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{postNormMean}
\alias{postNormMean}
\title{Mean of normalized power prior}
\usage{
postNormMean(dat, vardata, priormean, priorvar, alpha = 1)
}
\arguments{
\item{dat}{Data.}

\item{vardata}{Variance of the data.}

\item{priormean}{Mean parameter of initial normal prior. Defaults to
\code{0}.}

\item{priorvar}{Variance parameter of initial normal prior. Defaults to
\code{Inf} (uniform prior).}

\item{alpha}{Power parameter. Indicates to which power the likelihood of the
data is raised. Can be set to a number in [0, 1]. Defaults to \code{1}.}
}
\value{
Prior mean
}
\description{
This function computes the mean of a normalized power prior
    conditional on a fixed power parameter and an initial normal prior for
    the effect size.
}
\author{
Samuel Pawel
}
\keyword{internal}
