% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppRepPKG.R
\docType{package}
\name{ppRepPKG}
\alias{ppRepPKG}
\title{ppRep: Bayesian analysis of replication studies using power priors}
\description{
The ppRep package includes methods for Bayesian estimation and
    hypothesis testing of effect sizes based on original and replication
    study using the power prior framework.
}
\references{
Pawel, S., Aust, F., Held, L., and Wagenmakers, E.-J. (2022).
    Power Priors for Replication Studies. arXiv preprint.
    \doi{10.48550/arXiv.2207.14720}
}
